gnrc: mark static _event_loop function as unsafe
All `_event_loop` functions are unsafe because the IPC `msg_t` type
doesn't have a member for storing the size of the message. Therefore
`msg.content.ptr` is a `_Ptr<void>` instead of a `_Array_ptr<void>` type
with an associated bytes_count and thus the `_Dynamic_bounds_cast`
previously used in this function will always fail.

This has been fixed by using `_Assume_bounds_cast` instead and marking
the function as unchecked. An alternative solution would be modifying
the `msg_t` struct and adding a length parameter for the
`msg.content.ptr` pointer to it.