gnrc/netif: add bounds-safe interface for gnrc_netif_hdr_build
It would be preferable to add bounds-safe interface to all function from
this header, however, most static inline functions from this header
intentionally overflow the memory region for the `gnrc_netif_hdr_t` type
because it is assumed that the link layer address for the packet
associated with the header are put in memory directly after the
`gnrc_netif_hdr_t`.

Thus it's probably easier to add bounds-safe interfaces for functions
from this header file when we need them.